# public/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CrewAI Orchestrator API",
    "version": "1.0.0",
    "description": "Multi-Agent AI Orchestration Platform API",
    "contact": {
      "name": "API Support",
      "url": "https://github.com/yourusername/crewai-orchestrator"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    },
    {
      "url": "https://your-domain.com/api",
      "description": "Production server"
    }
  ],
  "paths": {
    "/agents": {
      "get": {
        "summary": "Get all agents",
        "description": "Retrieve a list of all AI agents",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "agents": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Agent"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create new agent",
        "description": "Create a new AI agent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Agent created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "agent": {
                      "$ref": "#/components/schemas/Agent"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Agent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "goal": {
            "type": "string"
          },
          "backstory": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "temperature": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "max_tokens": {
            "type": "integer"
          },
          "tools": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": ["active", "idle", "disabled"]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AgentInput": {
        "type": "object",
        "required": ["name", "role", "goal", "model"],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "role": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "goal": {
            "type": "string",
            "minLength": 10,
            "maxLength": 500
          },
          "backstory": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "temperature": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "default": 0.7
          },
          "maxTokens": {
            "type": "integer",
            "minimum": 100,
            "maximum": 8000,
            "default": 2000
          },
          "tools": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": ["active", "idle", "disabled"],
            "default": "active"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}